add some comments and move carousel into own folder within explore
